fix(ci): merge migration heads and keep proxy facade under budget
